requirements.txt
langchain
python-dotenv
ipykernel
langchain_community
pypdf
bs4
langchain-text-splitters
pymupdf
sentence_transformers
langchain_huggingface
faiss-cpu
chromadb
langchain_chroma
streamlit
langchain-openai
langchain_groq
fastapi
uvicorn
langserve
arxiv
wikipedia